+++ database/migrations/2021_03_14_031208_create_detail_transaksi_table.php
@@ -17,7 +17,7 @@ class CreateDetailTransaksiTable extends Migration
             $table->integer('transaksi_id');
             $table->integer('produk_id');
             $table->integer('qty');
-            $table->float('harga');
+            $table->bigInteger('harga');
         });
     }
